Add conda packaging
This commit adds conda packaging for recast.

One change to the CMakeLists.txt was necessary to make it build correctly:
- Add imgui dependency on libdl
  - Imgui.a contains a call to dlclose. It thus has to be linked to libdl.

The conda recipe pulls the most recent source code from the master
branch of recast3d.  Because of the afore-mentioned change to
CMakeLists.txt, the conda recipe will not work until this pull request
is merged into master.

----------------------------------------------------------------------

I have chosen to split the dependencies as follows:

Conda provides the following dependencies:
- OpenGL (required) development libraries and headers
- ZeroMQ (submodule)
- Eigen (submodule)

The following packages are built from source (submodules):
- glm (required)
- GLFW (submodule)
- Assimp (submodule)
- ImGui (submodule)
- TomoPackets (submodule)

This one I am actually not sure about:
- cppzmq (submodule)

The packages assimp, glm, and glfw are not provided by the defaults
channel in anaconda. Conda-forge does provide these packages, but has
a different packaging philosophy for opengl and x11 that I am not sure
is maximally compatible across all devices. In any case, building the
dependencies from source allows us to build the conda package without
adding "-c conda-forge".

NOTE: I have added numerous dependecies to so-called CDTs in the build
requirements. I am sure these are not all required. Nonetheless, they
do not add any weight to the resulting package and are fast to
download. So I have decided against checking each and every CDT
dependency to see if it can be removed.

----------------------------------------------------------------------

I have updated the tomopackets dependency to the master branch.

----------------------------------------------------------------------

PS. I sneaked in a small change to .gitignore:
- Ignore files ending in ~